requirements.txt
# Fixing sphinx version due to https://github.com/sphinx-doc/sphinx/issues/3976
setuptools==46.4.0
plyvel==1.2.0
ntplib==0.3.4
Twisted>=19.10.0
colorlog==3.1.0
simplejson==3.11.1
PyYAML==5.3.1
grpcio-tools>=1.9.0,<=1.27.2
grpcio>=1.9.0,<=1.27.2
google-api-python-client==1.8.3
service_identity==17.0.0
protobuf==3.6.0
pyopenssl==17.5.0
six==1.11
click==6.7
pyxmlib>=0.99.3,<1.1.0
pyqryptonight>=0.99.3,<1.1.0
pyqrandomx>=0.0.0,<1.0.0
Flask>=1.0.0
json-rpc==1.10.8
cryptography==2.3
mock==2.0.0
daemonize==2.4.7
